# server/doc/swagger.yaml
openapi: 3.0.0
info:
  contact:
    email: snowmerak@outlook.com
    name: snowmerak
  description: The ggeco API Documentation
  title: ggeco
  version: 1.0.0
paths:
...
  /api/badge/name:
    get:
      description: Get Badge data by badge name
      parameters:
      - in: query
        name: name
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BadgesGetBadgeResponse'
                type: array
          description: OK
      summary: Get Badge data
    post:
      description: Update Badge Name
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BadgesUpdateBadgeNameRequest'
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Update Badge Name
...
components:
  schemas:
...
    BadgesGetBadgeResponse:
      properties:
        id:
          type: string
        image:
          type: string
        name:
          type: string
        summary:
          type: string
      type: object
...
    BadgesUpdateBadgeNameRequest:
      properties:
        badge_id:
          type: string
        name:
          type: string
      required:
      - badge_id
      - name
      type: object
...